fix "Window to Screen" regression (#52)
 * This fixes the regression where "Window to Screen" shortcut fails
because of Krohnkite. This worked in some previous revisions

 * The cause of the problem was minimizeRest feature implemented
recently. Originally, Krohnkite arranged windows as soon as a window is
move to another screen, but the new feature delayed this re-arragement.
This caused KWin to fire subsequent `geometryChanged` signal, causing
`enforceSize` to be called to revert any changes to the window.

 * This patch both fixes the regression and keeps the feature, but might
face some race conditions, which is inevitable due to how KWin operates.
KWin emits `clientUnminimized` as soon as user gives the order, but the
unminimized window can get focus only after unminimize aninmation is
finished.