feat(wayland): :fire: remove qmlSetTimer
It does not work for some reason. Using timer is not a good idea and if
we a bound to do so, it is better to create destinct ones in the qml
itself instead of using arbitary number of ones with singleton.

This will potentially break the code in places, where timer was used. We
need to test it and provide distinct timer in QML, which are working.