chore: basic core plugin
This plugin will encapsulate all the logic of the window tiling
extension. It is intended to replace the legacy TypeScript logic,
to be more robust, testable and more feature rich.

The explanations on why TypeScript is not a way to go can be found in
the following section of the blog post:

https://genda.life/posts/bismuth-future/#motivation