chore: replace liberapay with open collective
PayPal is going to block Russian users. Because of that, I (gikari) will
no longer able to receive donations from people around the world.

To keep the ability for people to donate the open collective account was
created (https://opencollective.com/bismuth). It will be able to keep
the funds for the time being, even if they won't be available to me
in the near future.