feat: Add VerticalTileLayout
A Vertical tiling layout which is tailored for portrait orientated
monitors but might have use cases in landscape as well.

The intended behavior is for the master area to take a fixed (but
configurable via shortcut) percentage of the screen and for the
remaining panes to split the remaining space. Multiple masters are
supported and will share the master area equally.

Two masters, followed by a stack of two panels:
  ---------
 |         |
 |   M1    |
 |         |
 | ------- |
 |         |
 |   M2    |
 |         |
  ---------
 |    1    |
  ---------
 |    2    |
  ---------