Adjust KCM installation, clean up metadata
We do not want the KCM to be shown in the plugin selector, but only the
standalone KCM.
By not specifying the parent components and copying the KService, we do
not use the deprecated KCM loading and don't load a KCM at all.

While at it, clean up the metadata and install the KCM in the new
location.